Create upload page.
This commit creates a sample upload page using Django's ModelForm and
@login_required.

Also, login_register.css was moved to /static/app/form_page.css, for it
to serve as a generic css for pages whose main content is a form. The
specific login and register styles haven't been separated from the file
yet though.